#3 - Added RxKotlin

+++ app/src/main/java/com/leticiafernandes/letsmovie/infrastructure/data/entity/Movie.kt
@@ -11,7 +11,7 @@ class Movie(var id: Long, @SerializedName("vote_count") var voteCount: Int, var 
             var popularity: Double, @SerializedName("poster_path") var posterPath: String,
             @SerializedName("original_language") var originalLanguage: String,
             @SerializedName("original_title") var originalTitle: String,
-            @SerializedName("genre_ids") var genreIds: List<Int>,
+            @SerializedName("genre_ids") var genreIds: List<Long>,
             @SerializedName("backdrop_path") var backdropPath: String,
             var adult: Boolean, var overview: String,
             @SerializedName("release_date") var releaseDate: Date) {
